Create DowloadURLs-FFMpeg.vbs
Downloads a list of TV Show URLs contained in a text file. The name of the text file should be the season number (S01.txt, S02.txt, etc...). The episode numbers can be included in the text file at the beginning of each line and separated by a comma with the URL. (1,http://www.sourcesite.com/tvshowurl.m3u8)